Continuing adaptation to OOP
